feat(react): migrate next builders to devkit
