+++ e2e/workspace/src/create-nx-workspace.test.ts
@@ -68,6 +68,8 @@ describe('create-nx-workspace', () => {
       style: 'css',
       appName,
     });
+
+    expectNoAngularDevkit();
   });
 
   it('should be able to create an web-components workspace', () => {
