+++ packages/next/package.json
@@ -37,9 +37,7 @@
     "@nrwl/jest": "*",
     "@nrwl/linter": "*",
     "@nrwl/web": "*",
-    "@angular-devkit/architect": "~0.1102.0",
-    "@angular-devkit/schematics": "~11.2.0",
-    "@angular-devkit/core": "~11.2.0",
+    "@nrwl/workspace": "*",
     "@svgr/webpack": "^5.4.0",
     "chalk": "4.1.0",
     "copy-webpack-plugin": "6.0.3",
