+++ packages/next/src/executors/export/compat.ts
@@ -1,0 +1,5 @@
+import { convertNxExecutor } from '@nrwl/devkit';
+
+import exportExecutor from './export.impl';
+
+export default convertNxExecutor(exportExecutor);
