+++ packages/next/src/executors/export/schema.json
@@ -1,4 +1,5 @@
 {
+  "cli": "nx",
   "title": "Next Export",
   "description": "Export a Next.js app. The exported application is located at dist/$outputPath/exported.",
   "type": "object",
