+++ packages/next/src/executors/server/compat.ts
@@ -1,0 +1,5 @@
+import { convertNxExecutor } from '@nrwl/devkit';
+
+import serverExecutor from './server.impl';
+
+export default convertNxExecutor(serverExecutor);
