--- packages/next/src/utils/testing.ts
@@ -1,25 +1,0 @@
-import { Architect } from '@angular-devkit/architect';
-import { TestingArchitectHost } from '@angular-devkit/architect/testing';
-import { schema } from '@angular-devkit/core';
-import { MockBuilderContext } from '@nrwl/workspace/testing';
-import { join } from 'path';
-
-export async function getTestArchitect() {
-  const architectHost = new TestingArchitectHost('/root', '/root');
-  const registry = new schema.CoreSchemaRegistry();
-  registry.addPostTransform(schema.transforms.addUndefinedDefaults);
-
-  const architect = new Architect(architectHost, registry);
-
-  await architectHost.addBuilderFromPackage(join(__dirname, '../..'));
-
-  return [architect, architectHost] as [Architect, TestingArchitectHost];
-}
-
-export async function getMockContext() {
-  const [architect, architectHost] = await getTestArchitect();
-
-  const context = new MockBuilderContext(architect, architectHost);
-  await context.addBuilderFromPackage(join(__dirname, '../..'));
-  return context;
-}
